Refactor: custom resolvers for moderation
