+++ src/resolvers/moderation.js
@@ -1,0 +1,10 @@
+export default {
+  Mutation: {
+    disable: async (object, params, {user, driver}) => {
+      return true
+    },
+    enable: async (object, params, {user, driver}) => {
+      return true
+    }
+  }
+}
